Remove all files for the locally implemented SHA256 function
Solely rely on the openssh implementation of sha256 which is typically way faster due to binary files. This means that in order to compile, the files need to be installed on the system. This should not be a problem as a lot of distributions already have these files in their repositories